post

+++ app/Http/Controllers/TasksController.php
@@ -11,13 +11,22 @@ class TasksController extends Controller
     // getでtasks/にアクセスされた場合の「一覧表示処理」
     public function index()
     {
-        // タスク一覧を取得
-        $tasks = Task::paginate(25);
-
-        // タスク一覧ビューでそれを表示
-        return view('tasks.index', [
+        $data = [];
+        if (\Auth::check()) { // 認証済みの場合
+            // 認証済みユーザを取得
+            $user = \Auth::user();
+            // ユーザの投稿の一覧を作成日時の降順で取得
+            $tasks = $user->tasks()->orderBy('created_at', 'desc')->paginate(10);
+            
+            $data = [
+            'user' => $user,
             'tasks' => $tasks,
-        ]);
+            ];
+        }
+        
+        // タスク一覧ビューでそれを表示
+        return view('tasks.index', $data);
+        
     }
 
     // getでtasks/createにアクセスされた場合の「新規登録画面表示処理」
@@ -40,11 +49,12 @@ class TasksController extends Controller
             'status' => 'required|max:10',
         ]);
         
-        // タスクを作成
-        $task = new Task;
-        $task->content = $request->content;
-        $task->status = $request->status;
-        $task->save();
+        // 認証済みユーザ（閲覧者）の投稿として作成（リクエストされた値をもとに作成）
+        $request->user()->tasks()->create([
+            'content' => $request->content,
+            'status' => $request->status
+        ]);
+
 
         // トップページへリダイレクトさせる
         return redirect('/');
@@ -98,10 +108,13 @@ class TasksController extends Controller
     public function destroy($id)
     {
         // idの値でタスクを検索して取得
-        $task = Task::findOrFail($id);
-        // タスクを削除
-        $task->delete();
-
+        $task = \App\Task::findOrFail($id);
+        
+        // 認証済みユーザ（閲覧者）がその投稿の所有者である場合は、投稿を削除
+        if (\Auth::id() === $task->user_id) {
+            $task->delete();
+        }
+        
         // トップページへリダイレクトさせる
         return redirect('/');
     }
